updating download script
Former-commit-id: 6a6e6e16c1e4d2b2e358f1c82d1a996dc930a41b [formerly 6a6e6e16c1e4d2b2e358f1c82d1a996dc930a41b [formerly 8211d484882772e4de0b2b9e68debd3d68f1e449]]
Former-commit-id: 788d515139840e8ab6998088a680bf7ebd0ec50b
Former-commit-id: 86fca4fb6a5baecfadea91d17131ca313c8d4ead

+++ data/download_data.sh
@@ -28,8 +28,19 @@ wget https://aka.ms/alfworld/mrcnn.pth
 mkdir -p $ALFRED_ROOT/agents/detector/models/
 mv mrcnn.pth $ALFRED_ROOT/agents/detector/models/
 
-# Seq2Seq training files
-rm seq2seq_data.zip
-wget https://aka.ms/seq2seq_data.zip
-unzip seq2seq_data.zip
-rm seq2seq_data.zip
+# Pre-trained BUTLER agents
+if [ "$1" = "extra" ]
+then
+    rm alfworld_pretrained_checkpoints.zip
+    wget https://aka.ms/alfworld_pretrained_checkpoints.zip
+    unzip alfworld_pretrained_checkpoints.zip
+    mkdir -p $ALFRED_ROOT/agents/runs/
+    mv pretrained_checkpoints/* $ALFRED_ROOT/agents/runs/
+    rm alfworld_pretrained_checkpoints.zip
+
+    # Seq2Seq training files
+    rm seq2seq_data.zip
+    wget https://aka.ms/seq2seq_data.zip
+    unzip seq2seq_data.zip
+    rm seq2seq_data.zip
+fi
